Add unit testing

+++ src/main/java/com/ss/training/utopia/agent/dao/AirportDAO.java
@@ -10,5 +10,5 @@ import org.springframework.stereotype.Repository;
  */
 @Repository
 public interface AirportDAO extends JpaRepository<Airport, Long> {
-    Airport findByAirportId(Integer id);
+    Airport findByAirportId(Long id);
 }
